+++ src/v1_omp.c
@@ -33,7 +33,7 @@ int* CuthillMckee(int* matrix) {
   omp_init_lock(&writelock);
 
   // init notVisited array
-  for (size_t i = 0; i < SIZE; i++) {
+  for (int i = 0; i < SIZE; i++) {
     *(notVisited+i) = 1;
   }
 
@@ -74,7 +74,7 @@ int* CuthillMckee(int* matrix) {
       sortByDegree(neighbors, degrees, neighborsCounter);
 
       // add the sorted neighbor in Q
-      for (size_t i = 0; i < neighborsCounter; i++) {
+      for (int i = 0; i < neighborsCounter; i++) {
         queueAdd(Q, neighbors[i]);
       }
 
@@ -129,8 +129,8 @@ int main(int argc, char const *argv[]) {
   init_matrix(matrix, SIZE, MODE, SPARSITY);
 
   // printf("MATRIX:\n");
-  // for (size_t i = 0; i < SIZE; i++) {
-  //   for (size_t j = 0; j < SIZE; j++) {
+  // for (int i = 0; i < SIZE; i++) {
+  //   for (int j = 0; j < SIZE; j++) {
   //     printf("%d ", *(matrix+i*SIZE+j));
   //   }
   //   printf("\n");
@@ -160,7 +160,7 @@ int main(int argc, char const *argv[]) {
 
 
 
-  // for (size_t i = 0; i < SIZE; i++) {
+  // for (int i = 0; i < SIZE; i++) {
   //   printf("%d, ", R[i]);
   // }
