+++ src/v3_omp.c
@@ -33,7 +33,7 @@ int* CuthillMckee(int* matrix) {
   omp_set_dynamic(0);
   omp_set_nested(1);
 
-  for (size_t i = 0; i < SIZE; i++) {
+  for (int i = 0; i < SIZE; i++) {
     *(notVisited+i) = 1;
   }
 
@@ -71,7 +71,7 @@ int* CuthillMckee(int* matrix) {
         neighborsCounter[tid]=0;
 
         if(threads_using > MAX_THREADS/2) {
-          for (size_t i = 0; i < SIZE; i++) {
+          for (int i = 0; i < SIZE; i++) {
             if(i != currentIndex[tid] && *(matrix+(currentIndex[tid])*SIZE+i)==1 && notVisited[i]==1) {
               neighbors[tid][neighborsCounter[tid]] = i;
               neighborsCounter[tid]++;
@@ -145,8 +145,8 @@ int main(int argc, char const *argv[]) {
   init_matrix(matrix, SIZE, MODE, SPARSITY);
 
   // printf("MATRIX:\n");
-  // for (size_t i = 0; i < SIZE; i++) {
-  //   for (size_t j = 0; j < SIZE; j++) {
+  // for (int i = 0; i < SIZE; i++) {
+  //   for (int j = 0; j < SIZE; j++) {
   //     printf("%d ", *(matrix+i*SIZE+j));
   //   }
   //   printf("\n");
@@ -173,7 +173,7 @@ int main(int argc, char const *argv[]) {
 
 
 
-  // for (size_t i = 0; i < SIZE; i++) {
+  // for (int i = 0; i < SIZE; i++) {
   //   printf("%d, ", R[i]);
   // }
